refactor: Moved AppLayout to it's own file
